+++ README.md
@@ -10,20 +10,7 @@
 `api` contains classes available on all platforms. It defines the protocol between Lunar Client and servers / proxies.
 
 There are modules for different implementations of the API that are also plugins for their platform.
-These modules contain source sets for NMS and Mixin code.
-
-## Building
-
-Run Spigot BuildTools to get a copy of the latest Minecraft Spigot in your local maven installation.
-(This may take a while.)
-
-```shell
-mkdir workdir && cd workdir
-
-curl -o BuildTools.jar https://hub.spigotmc.org/jenkins/job/BuildTools/lastSuccessfulBuild/artifact/target/BuildTools.jar
-export MAVEN_OPTS="-Xmx2G" java -Xmx2G -jar BuildTools.jar --compile craftbukkit --rev 1.19
-export MAVEN_OPTS="-Xmx2G" java -Xmx2G -jar BuildTools.jar --compile craftbukkit --rev 1.18.2
-```
+These modules may contain subprojects for NMS and Mixin code as well.
 
 ### Bukkit
 
@@ -31,16 +18,14 @@ The Bukkit implementation is split into an Ignite mod and a Bukkit plugin.
 
 ```shell
 # Build jar of the Bukkit plugin
-./gradlew :apollo-bukkit:shadowJar
+./gradlew :apollo-bukkit:apollo-bukkit-common:build
 
 # Build jar of the 1.1X Ignite mod
-./gradlew :apollo-bukkit:jar_v1_18
-./gradlew :apollo-bukkit:jar_v1_19
+./gradlew :apollo-bukkit:apollo-bukkit-v1_18:build
+./gradlew :apollo-bukkit:apollo-bukkit-v1_19:build
 
-# Jars are stored in here
-ls bukkit/build/
 # ignite mods:
-jar -tf bukkit/build/libs/apollo-bukkit-v1_18.jar
+jar -tf bukkit/v1_18/build/libs/apollo-bukkit-v1_18.jar
 # bukkit plugin:
-jar -tf bukkit/build/libs/apollo-bukkit-all.jar
+jar -tf bukkit/common/build/libs/apollo-bukkit-common.jar
 ```
