+++ bukkit/common/src/main/java/com/moonsworth/apollo/impl/bukkit/listener/RegenListener.java
