+++ bukkit/common/src/main/resources/config.yml
