--- bukkit/src/craftbukkit_v1_18/java/com/moonsworth/apollo/impl/bukkit/v1_18/NmsTest.java
@@ -1,13 +1,0 @@
-package com.moonsworth.apollo.impl.bukkit.v1_18;
-
-import net.minecraft.world.entity.EntityLiving;
-import org.bukkit.craftbukkit.v1_18_R2.CraftChunk;
-
-public class NmsTest {
-
-    public static void doItCompile() {
-        var c = EntityLiving.class;
-        var craftChunkClass = CraftChunk.class;
-    }
-
-}
