--- bukkit/src/craftbukkit_v1_18/resources/mixins.apollo.v1_18.json
@@ -1,16 +1,0 @@
-{
-  "minVersion": "0.8",
-  "package": "com.moonsworth.apollo.impl.bukkit.v1_18.mixin",
-  "target": "CLIENT",
-  "required": true,
-  "compatibilityLevel": "JAVA_11",
-  "mixins": [
-    "EntityHumanMixin_v1_18",
-    "EntityLivingMixin_v1_18",
-    "EntityPotionMixin_v1_18",
-    "PlayerConnectionMixin_v1_18"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
