--- bukkit/src/craftbukkit_v1_19/java/com/moonsworth/apollo/impl/bukkit/v1_19/NmsTest.java
@@ -1,13 +1,0 @@
-package com.moonsworth.apollo.impl.bukkit.v1_19;
-
-import net.minecraft.world.entity.EntityLiving;
-import org.bukkit.craftbukkit.v1_19_R1.CraftChunk;
-
-public class NmsTest {
-
-    public static void doItCompile() {
-        var c = EntityLiving.class;
-        var craftChunkClass = CraftChunk.class;
-    }
-
-}
