--- bukkit/src/craftbukkit_v1_19/resources/mixins.apollo.v1_19.json
@@ -1,15 +1,0 @@
-{
-  "minVersion": "0.8",
-  "package": "com.moonsworth.apollo.impl.bukkit.v1_19.mixin",
-  "target": "CLIENT",
-  "required": true,
-  "compatibilityLevel": "JAVA_11",
-  "mixins": [
-    "EntityHumanMixin_v1_19",
-    "EntityLivingMixin_v1_19",
-    "PlayerConnectionMixin_v1_18"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
