+++ bukkit/v1_18/src/main/java/com/moonsworth/apollo/impl/bukkit/v1_18/ApolloIgniteMod_v1_18.java
