+++ bukkit/v1_18/src/main/resources/ignite.mod.json
