+++ bukkit/v1_19/src/main/java/com/moonsworth/apollo/impl/bukkit/v1_19/NmsTest.java
@@ -1,0 +1,13 @@
+package com.moonsworth.apollo.impl.bukkit.v1_19;
+
+import net.minecraft.world.entity.LivingEntity;
+import org.bukkit.craftbukkit.v1_19_R2.CraftChunk;
+
+public class NmsTest {
+
+    public static void doItCompile() {
+        var c = LivingEntity.class;
+        var craftChunkClass = CraftChunk.class;
+    }
+
+}
