+++ bukkit/v1_19/src/main/resources/ignite.mod.json
