+++ settings.gradle.kts
@@ -1,11 +1,19 @@
 rootProject.name = "apollo"
-includeBuild("toolchain")
-include("api")
-include("bukkit")
-include("bungee")
-include("velocity")
 
-findProject(":api")?.name = "apollo-api"
-findProject(":bukkit")?.name = "apollo-bukkit"
-findProject(":bungee")?.name = "apollo-bungee"
-findProject(":velocity")?.name = "apollo-velocity"
+val mcVersion = startParameter.projectProperties["mc"]
+val loadAllVersions = mcVersion == null
+
+val modules = listOfNotNull(
+    "api",
+    "bukkit",
+    "bukkit:common",
+    if (loadAllVersions || mcVersion == "v1_18") "bukkit:v1_18" else null,
+    if (loadAllVersions || mcVersion == "v1_19") "bukkit:v1_19" else null,
+    if (loadAllVersions) "bungee" else null,
+    if (loadAllVersions) "velocity" else null
+)
+
+includeBuild("toolchain")
+include(modules)
+
+modules.forEach { findProject(":$it")?.name = "apollo-${it.replace(':', '-')}" }
